pyproject.toml
[project]
name = "papersys"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datauri>=1.0.0",
    "jupyter>=1.1.1",
    "huggingface-hub>=0.24.6",
    "litellm>=1.79.0",
    "loguru>=0.7.3",
    "marker-pdf>=1.10.1",
    "mistralai>=1.9.11",
    "markdown-it-py[linkify,plugins]>=3.0.0",
    "mdit-py-plugins>=0.4.0",
    "polars>=1.34.0",
    "python-dotenv>=1.1.1",
    "tqdm>=4.67.1",
    "typer>=0.20.0",
    "ultimate-notion>=0.9.5",
    "notion-client>=2.5.0",
]

[project.scripts]
papersys = "papersys.cli:app"

[tool.uv]
package = true

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["papersys"]
